// src/modules/resources/virtual-transactions/dto/general.dto.ts
export enum VirtualTransactionStatus {
  PENDING = 'PENDING',
  INITIAL = 'INITIAL',
  PROCESSING = 'PROCESSING',
  SUCCEED = 'SUCCEED',
  CANCELED = 'CANCELED',
  REVIEW = 'REVIEW',
  REFUNDING = 'REFUNDING',
  REFUNDED = 'REFUNDED',
  WAIT_INSTALLMENT = 'WAIT_INSTALLMENT',
  REFUNDED_A_PART = 'REFUNDED_A_PART'
}

export enum VirtualServiceCode {
  COLLECTCASH = 'COLLECTCASH',
  PAYCASH = 'PAYCASH'
}